services/clustering/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
umap-learn==0.5.6
hdbscan==0.8.40
numpy==1.26.4
scikit-learn==1.5.2
